Fix error message for posix_spawn
Set the USEVFORK flag for posix_spawn. This flag tells posix_spawn to
use vfork instead of fork, which suspends the parent process until the
child calls _exit or _execve and allows the child process to share the
same memory with its parent until execve is called by the child. Hence,
execve can set errno in the child process, and when the parent resumes
errno will be set if execve failed.