Fix posix_spawn error detection
When `posix_spawnp` is executed with a command that is in the PATH it
sets `errno` to ENOENT even if the command is found. Hence, there is no
way to determine if `posix_spawnp` succeeded when executing a command in
the PATH. As an alternative, always use fork with the self-pipe trick to
detect an error in `execvp`. This forces the use of the internally
defined version of `posix_spawn`, on all unix variants, to prevent
linking with any version of `posix_spawn` installed on the system.